Incluindo filtro de valor do DTO de contas pagas

+++ src/main/java/br/com/etec/exerc/trabalho/model/PayAccount.java
@@ -27,6 +27,6 @@ public class PayAccount {
     private Client client;
 
     @Column
-    private BigDecimal value;
+    private Double value;
 
 }
